On Writing
I checked out this book from the library on a complete whim last summer. I _loved_ it. There is a lot of overlap between this and Elizabeth Gilbert's Big Magic (though Stephen King's muse is a surly old guy with a cigar, and Elizabeth's is more of a sexier, mystical creature).