How to Do Nothing
This book was _delicious_, which is a strange way to describe a book, but it was one of those that left that "writer glow" on me. That feeling of, "whoa, I want to be able to write like _this_."

Jenny adds a couple distinct things to this outline. First, the section on "Paying Attention." I think that was one line before Jenny arrived. Also, curation as a creative act. Jenny takes the work of others and expands it. Connects it. Makes it richer than it was before. I read both How to Do Nothing and Jia Tolentino's Trick Mirror around the same time, and I loved how they both riffed off one another. Taking an idea here and an idea there and taking it a step or two further. There's this pressure associated with being original — which is why I'm enamored with the notion of the genius as separate entity that both Gilbert and King describe — but the best work is built on that of others (really, this applies to _all work_).