Creativity
This book — and Csikszentmihalyi's work in general — is as good an example as any of what I'd like to spend my life doing. Research, writing, and synthesizing how we can lead more interesting, engaged, creative lives.

I love this book. It's one I will read and re-read throughout my life. I've underlined and dog-eared it, used it as a notepad to jot down restaurants to eat at and places to see while travelling. It's been good to me.

This book is also loooooong, and as I added in my highlights and notes, I ended up inserting a number of choice highlights from some of the other reading I was doing on a rainy Saturday afternoon indoors, when the world around me was starting to fall into pieces.

That reading largely centered on a few Jenny Odell interviews and Heather Havrilesky's "Ask Polly" column, which is fucking delightful, and if you've been a dunce and haven't read it yet like me, I encourage you to stop what you're doing and remedy your situation post-hence.