How to Have a Good Day
I remember when I first read this book. It was the same reaction I had after reading Creativity and Flow: "This is what I want to be able to do!"

I don't think this book has received nearly as much attention as it deserves. Deep Work, for instance, was talked about in all sorts of tech circles when it first came out, and Cal Newport was _heavily_ influenced by Caroline Webb's work. It's referenced throughout Deep Work, and the concept for batching your day is basically a more actionable, detailed version of what's talked about in How to Have a Good Day.

There's a lot from this book that I'll have to capture in a future guide. Something that's more work related — how to work with others, how to ask great questions at work — something that's along those lines.