Rewrite-2
I must have been feeling really sentimental this morning, because when I came back to this in the evening I was like...um, what is this. Writing's hard, man. But it's even harder when you're writing for an audience that doesn't exist. I decided not to adjust my tone or voice, or try to speak to anyone in particular beyond myself. I decided to write like I would in my journal, and see what comes of it.